Some refactoring and reformatting
