+++ lib/main.dart
@@ -1,8 +1,7 @@
-
 import 'package:flutter/material.dart';
-import 'package:homeworkplanner/pages/plannerpage.dart';
-import 'package:homeworkplanner/tasksystem/savefile.dart';
-import 'package:homeworkplanner/tasksystem/taskhost.dart';
+import 'package:homeworkplanner/ui/main_page.dart';
+import 'package:homeworkplanner/models/tasksystem/savefile.dart';
+import 'package:homeworkplanner/models/tasksystem/taskhost.dart';
 
 void main() {
   runApp(const MainApp());
@@ -17,12 +16,13 @@ class MainApp extends StatefulWidget {
 
 class _MainAppState extends State<MainApp> {
   TaskHost host = TaskHost(saveFile: SaveFile());
-  
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-        theme: ThemeData(primarySwatch: Colors.blue), 
-        home: PlannerPage(host: host,)
-    );
-  } 
+        theme: ThemeData(primarySwatch: Colors.blue),
+        home: MainPage(
+          host: host,
+        ));
+  }
 }
