+++ lib/models/lists/schedulelist.dart
@@ -1,5 +1,5 @@
 // ignore_for_file: non_constant_identifier_names
- 
+
 import 'package:homeworkplanner/models/main/schedule.dart';
 import 'package:json_annotation/json_annotation.dart';
 
@@ -17,6 +17,6 @@ class ScheduleList {
   }
 
   factory ScheduleList.fromJson(Map<String, dynamic> json) => _$ScheduleListFromJson(json);
-  
+
   Map<String, dynamic> toJson() => _$ScheduleListToJson(this);
-} 
+}
