+++ lib/models/main/subject.dart
@@ -25,6 +25,6 @@ class Subject {
   }
 
   factory Subject.fromJson(Map<String, dynamic> json) => _$SubjectFromJson(json);
-  
+
   Map<String, dynamic> toJson() => _$SubjectToJson(this);
 }
