+++ lib/models/tasksystem/savefile.dart
@@ -4,7 +4,7 @@ import 'package:homeworkplanner/models/lists/schedulelist.dart';
 import 'package:homeworkplanner/models/lists/subjectlist.dart';
 import 'package:homeworkplanner/models/lists/tasklist.dart';
 import 'package:homeworkplanner/models/main/daynote.dart';
-import 'package:homeworkplanner/tasksystem/savesettings.dart';
+import 'package:homeworkplanner/models/tasksystem/savesettings.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'savefile.g.dart';
@@ -21,6 +21,6 @@ class SaveFile {
   factory SaveFile.fromJson(Map<String, dynamic> json) => _$SaveFileFromJson(json);
 
   Map<String, dynamic> toJson() => _$SaveFileToJson(this);
-  
+
   SaveFile();
 }
