+++ lib/ui/taskpage.dart
@@ -3,31 +3,27 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:homeworkplanner/models/main/subject.dart';
 import 'package:homeworkplanner/models/main/task.dart';
-import '../tasksystem/taskhost.dart';
+import '../models/tasksystem/taskhost.dart';
 
-class TaskPage extends StatefulWidget {
+class TaskEditorPage extends StatefulWidget {
   Task task;
   TaskHost host;
 
-  TaskPage({super.key, required this.task, required this.host});
+  TaskEditorPage({super.key, required this.task, required this.host});
 
   @override
-  State<TaskPage> createState() => _TaskPageState(task: task, host: host);
+  State<TaskEditorPage> createState() => _TaskEditorPageState(task: task, host: host);
 }
 
-class _TaskPageState extends State<TaskPage> {
+class _TaskEditorPageState extends State<TaskEditorPage> {
   Task task;
   TaskHost host;
 
-  _TaskPageState({required this.task, required this.host});
+  _TaskEditorPageState({required this.task, required this.host});
 
   @override
   Widget build(BuildContext context) {
-    TaskPageBuilder builder = TaskPageBuilder(
-        onTaskCompleted: taskCompleted,
-        onTaskMarkedImportant: taskMarkedImportant,
-        setState: setState,
-        host: host);
+    TaskEditor builder = TaskEditor(onTaskCompleted: taskCompleted, onTaskMarkedImportant: taskMarkedImportant, setState: setState, host: host);
 
     return Scaffold(
         appBar: AppBar(
@@ -37,15 +33,14 @@ class _TaskPageState extends State<TaskPage> {
                 onPressed: () {
                   host.saveFile.Tasks.Items.remove(task);
                   Navigator.pop(context);
-                  ScaffoldMessenger.of(context)
-                      .showSnackBar(const SnackBar(content: Text('Task deleted')));
+                  ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text('Task deleted')));
                   ScaffoldMessenger.of(context).setState(() {});
                 },
                 icon: const Icon(Icons.delete))
           ],
         ),
         body: ListView(
-          children: builder.buildPageContent(task),
+          children: builder.build(task),
         ));
   }
 
@@ -55,38 +50,29 @@ class _TaskPageState extends State<TaskPage> {
     });
   }
 
-  void taskMarkedImportant(
-      Task task, bool value, Function(Function()) setState) {
+  void taskMarkedImportant(Task task, bool value, Function(Function()) setState) {
     setState(() {
       task.IsImportant = value;
     });
   }
 }
 
-class TaskPageBuilder {
+class TaskEditor {
   final Function(Task, bool, Function(Function())) onTaskCompleted;
   final Function(Task, bool, Function(Function())) onTaskMarkedImportant;
   final Function(Function()) setState;
   TaskHost? host;
 
-  TaskPageBuilder(
-      {required this.onTaskCompleted,
-      required this.onTaskMarkedImportant,
-      required this.setState,
-      required this.host});
+  TaskEditor({required this.onTaskCompleted, required this.onTaskMarkedImportant, required this.setState, required this.host});
 
-  List<Widget> buildPageContent(Task task) {
-    // String desc = "";
-    // for (var i = 0; i < task.Description.length; i++) {
-    //   desc += task.Description[i];
-    //   if (i != task.Description.length - 1) {
-    //     desc += "\n";
-    //   }
-    // }
+  List<Widget> build(Task task) {
     Subject noSubject = Subject.getNoSubject();
     Subject? selectedSubject = task.SubjectID == -1 ? noSubject : host!.getSubjectById(task.SubjectID);
     List<DropdownMenuItem<Subject>>? subjectWidgets = List.empty(growable: true);
-    subjectWidgets.add(DropdownMenuItem(child: Text(noSubject.SubjectName), value: noSubject,));
+    subjectWidgets.add(DropdownMenuItem(
+      child: Text(noSubject.SubjectName),
+      value: noSubject,
+    ));
     List<DropdownMenuItem<Subject>>? subjectWidgetsObtained = host?.saveFile.Subjects.Items.map<DropdownMenuItem<Subject>>(
       (e) {
         return DropdownMenuItem<Subject>(
@@ -119,10 +105,7 @@ class TaskPageBuilder {
       Padding(
         padding: const EdgeInsets.fromLTRB(16, 8, 16, 8),
         child: DropdownButtonFormField(
-          decoration: const InputDecoration(
-              icon: Icon(Icons.assignment_ind_outlined),
-              border: OutlineInputBorder(),
-              labelText: 'Subject'),
+          decoration: const InputDecoration(icon: Icon(Icons.assignment_ind_outlined), border: OutlineInputBorder(), labelText: 'Subject'),
           items: subjectWidgets,
           value: selectedSubject,
           onChanged: (value) {
@@ -140,10 +123,7 @@ class TaskPageBuilder {
       Padding(
         padding: const EdgeInsets.fromLTRB(16, 8, 16, 8),
         child: TextFormField(
-          decoration: const InputDecoration(
-              icon: Icon(Icons.description_outlined),
-              border: OutlineInputBorder(),
-              labelText: 'Description'),
+          decoration: const InputDecoration(icon: Icon(Icons.description_outlined), border: OutlineInputBorder(), labelText: 'Description'),
           maxLines: Platform.isAndroid ? 15 : 5,
           keyboardType: TextInputType.multiline,
           initialValue: task.Description,
